Additional changes to `EmbeddingManager.forward`.
Fewer copies to the device and some gradients are preserved by using the token already in `tokenized_text` instead of the unrelated `placeholder_token`.  This can be advantageous by allowing the embedding to learn using nuances like token ordering, as `tokenized_text` and `embedded_text` are related to each other.